Create Testing_code.py
This is the code to load the deep learning model into the Colab sheet and evaluate it on any set of images different from the initial training and testing set used to train and test the model. This code can be used at laboratories, as the scientists there can run their satellite images  through this code to get the classifications.